+++ R/listr.R
@@ -1,5 +1,5 @@
 #' @title
-#' Extended List Functionalities
+#' Extended List Functionality
 #'
 #' @description
 #' The package provides extended functionality for working with lists
@@ -24,6 +24,12 @@
 #'    \item{\code{\link[listr]{getByName}}: }{
 #'
 #'      Retrieve values by \code{name} as returned by 
+#'      \code{\link[listr]{getStructure}}
+#'    }
+#'    \item{\code{\link[listr]{getByPath}}: }{
+#'
+#'      Retrieve values by name-based (field \code{path}) or position-based 
+#'      paths (field \code{ppath}) as returned by 
 #'      \code{\link[listr]{getStructure}}
 #'    }
 #' }