+++ inst/examples/getByLevel.r
@@ -1,7 +1,8 @@
 \dontrun{
 
 input <- list(
-  x1 = list(x11 = list(x111 = 1, x112 = 1), x12 = list(x121 = 1, x122 = 1)),
+  x1 = list(x11 = list(x111 = list(x1111 = 1, x1112 = 2), x112 = 1), 
+    x12 = list(x121 = 1, x122 = 1)),
   x2 = list(x21 = "x21"),
   x3 = list("x31"),
   x4 = "x4",
@@ -17,14 +18,24 @@ input <- list(
   data.frame(x = 1:3, y = 1:3)
 )
 
-## Type 1 (top-level branches) //
-getByType(input, type = 1)
-## --> identical to `input`
+## Inspect structure //
+(struc <- getStructure(input))
 
-## Type 2 (intermediate-level branches) //
-getByType(input, type = 2)
+getByLevel(input, value = 1)
+getByLevel(input, value = 1, resolve = TRUE)
 
-## Type 3 (leafs) //
-getByType(input, type = 3)
+getByLevel(input, value = 2)
+getByLevel(input, value = 2, resolve = TRUE)
+
+getByLevel(input, value = 3)
+getByLevel(input, value = 3, resolve = TRUE)
+
+getByLevel(input, value = 4)
+getByLevel(input, value = 4, resolve = TRUE)
+
+## Condition handling //
+getByLevel(input, value = 5)
+try(getByLevel(input, value = 5, strict = 1))
+try(getByLevel(input, value = 5, strict = 2))
 
 }