+++ tests/testthat/test-getByLevel.r
@@ -2,7 +2,7 @@
 context("getByLevel/list")
 ##------------------------------------------------------------------------------
 
-test_that("getByLevel/list/level = 1", {
+test_that("getByLevel/list/value = 1", {
   
   input <- list(
     x1 = list(x11 = list(x111 = 1, x112 = 1), x12 = list(x121 = 1, x122 = 1)),
@@ -20,14 +20,16 @@ test_that("getByLevel/list/level = 1", {
     new.env(),
     data.frame(x = 1:3, y = 1:3)
   )
-  expect_is(res <- getByLevel(input, level = 1), "list")
+  expect_is(res <- getByLevel(input, value = 1), "list")
+  expect_identical(res, structure(input, names = rep("1", length(input))))
+  expect_is(res <- getByLevel(input, value = 1, resolve = TRUE), "list")
   expect_identical(res, input)
   
-  expect_is(res <- getByLevel(input, level = 1, parent = TRUE), "list")
+  expect_is(res <- getByLevel(input, value = 1, outer = TRUE), "list")
   
 })
 
-test_that("getByLevel/list/level = 2", {
+test_that("getByLevel/list/value = 2", {
   
   input <- list(
     x1 = list(x11 = list(x111 = 1, x112 = 1), x12 = list(x121 = 1, x122 = 1)),
@@ -45,12 +47,12 @@ test_that("getByLevel/list/level = 2", {
     new.env(),
     data.frame(x = 1:3, y = 1:3)
   )
-  expect_is(res <- getByLevel(input, level = 2), "list")
-  expect_is(res <- getByLevel(input, level = 2, parent = TRUE), "list")
+  expect_is(res <- getByLevel(input, value = 2), "list")
+  expect_is(res <- getByLevel(input, value = 2, outer = TRUE), "list")
   
 })
 
-test_that("getByLevel/list/level = 3", {
+test_that("getByLevel/list/value = 3", {
   
   input <- list(
     x1 = list(x11 = list(x111 = 1, x112 = 1), x12 = list(x121 = 1, x122 = 1)),
@@ -68,8 +70,8 @@ test_that("getByLevel/list/level = 3", {
     new.env(),
     data.frame(x = 1:3, y = 1:3)
   )
-  expect_is(res <- getByLevel(input, level = 3), "list")
-  expect_is(res <- getByLevel(input, level = 3, parent = TRUE), "list")
+  expect_is(res <- getByLevel(input, value = 3), "list")
+  expect_is(res <- getByLevel(input, value = 3, outer = TRUE), "list")
 #   getStructure(input)
   
 })
@@ -77,8 +79,8 @@ test_that("getByLevel/list/level = 3", {
 test_that("getByLevel/list/invalid", {
   
   input <- list(a = 1)
-  expect_identical(getByLevel(input, level = 2), list())
-  expect_warning(getByLevel(input, level = 2, strict = 1))
-  expect_error(getByLevel(input, level = 2, strict = 2))
+  expect_identical(getByLevel(input, value = 2), list())
+  expect_warning(getByLevel(input, value = 2, strict = 1))
+  expect_error(getByLevel(input, value = 2, strict = 2))
   
 })