+++ NEWS.md
@@ -1,3 +1,41 @@
+# CHANGES IN listr VERSION 0.2
+
+## NEW FEATURES
+
+- argument `resolve_by`:
+  Allows to choose the actual field on whoses basis name resolving/substitution is carried out (affects `getBy()` and all of its convencience wrappers)
+
+## BUG FIXES
+
+- couple of minor bug fixes w.r.t. names
+
+## MAJOR CHANGES
+
+- A: `getBy()`
+  Generic way to retrieve values by structure fields as returned by `getStructure()`. Other retrieval functions were refactored to be wrappers around `getBy()`
+
+## MINOR CHANGES
+
+## MISC
+
+-----
+
+# CHANGES IN listr VERSION 0.1.2
+
+## NEW FEATURES
+
+## BUG FIXES
+
+## MAJOR CHANGES
+
+## MINOR CHANGES
+
+## MISC
+
+- failed to bump version on GitHub
+
+-----
+
 # CHANGES IN listr VERSION 0.1.1
 
 ## NEW FEATURES